Add `ignore_exit_code` option
Godot 4.1.x has exit code 1 on MacOS currently. That causes the import step to fail.
This provides a workaround until this is fixed, and if it ever happens again. There may also
be other scenarios where it's better to keep going, such as if an export preset fails, to continue
to try the next export preset.

See https://github.com/godotengine/godot/issues/83449